+++ CHANGELOG.md
@@ -5,6 +5,8 @@
   * Add Email Address to User Hook Attributes.
   * Support managing group/project push rules.
   * Add `api::projects::merge_requests::MergeRequestDiffs` endpoint.
+  * `api::paged` iterators can be turned into async `Stream`s using the
+    `into_async()` method.
 
 ## Fixes